backend1/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
numpy==1.26.4
